upgraded to spring 1.2.1
git-svn-id: https://spring-rich-c.svn.sourceforge.net/svnroot/spring-rich-c/trunk/spring-richclient@658 817809c7-9d0e-0410-b92d-a7ac8b8adc13

+++ samples/petclinic/src/org/springframework/richclient/samples/petclinic/ui/OwnerManagerView.java
@@ -245,7 +245,7 @@ public class OwnerManagerView extends AbstractView implements ApplicationListene
     }
 
     private Owner reloadSelectedOwner() {
-        int ownerId = getSelectedOwner().getId();
+        int ownerId = getSelectedOwner().getId().intValue();
         return clinic.loadOwner(ownerId);
     }
